Bus: More restructure, new bus implementation
I've done a bit more restructuring of the project and have written
a new bus implementation `NesVectorBus` that should be better in
every way than the previous `NesMultiMapBus`.

+++ src/NesMultiMapBus.cpp
@@ -1,4 +1,4 @@
-#include "fmt/core.h"
+#include "fmt/printf.h"
 
 #include "NesMultiMapBus.h"
 
@@ -107,7 +107,7 @@ uint8_t NesMultiMapBus::read(uint16_t address, bool log, bool readOnly) {
 }
 
 void NesMultiMapBus::dump_memory(const char* filePath,
-	size_t startAddress, size_t endAddress) {
+	uint16_t startAddress, uint16_t endAddress) {
 
 	m_memDumpFile.open(filePath, std::ofstream::out);
